renamed comunicator to communicator & fixed problem with multiple connections
