+++ meter_solutions/apps/communicator/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from apps.comunicator.views import turn_on_view, turn_off_view
+from apps.communicator.views import turn_on_view, turn_off_view
 
 urlpatterns = [
     path('turn_off/<device_id>', turn_off_view),
