+++ meter_solutions/clock.py
@@ -1,3 +1,3 @@
-from apps.comunicator.services.receiver import background_receive
+from apps.communicator.services.receiver import background_receive
 
 background_receive()
